fix(plookup): fixed lookup vector: t must be ordered
