+++ ecc/bn254/fr/plookup/plookup_test.go
@@ -9,13 +9,13 @@ import (
 
 func TestH(t *testing.T) {
 
-	lookup := make(Table, 8)
-	f := make(Table, 7)
+	lookupVector := make(Table, 8)
+	fvector := make(Table, 7)
 	for i := 0; i < 8; i++ {
-		lookup[i].SetUint64(uint64(2 * i))
+		lookupVector[i].SetUint64(uint64(2 * i))
 	}
 	for i := 0; i < 7; i++ {
-		f[i].Set(&lookup[(4*i+1)%8])
+		fvector[i].Set(&lookupVector[(4*i+1)%8])
 	}
 
 	srs, err := kzg.NewSRS(64, big.NewInt(13))
@@ -23,9 +23,9 @@ func TestH(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	// correct proof
+	// correct proof vector
 	{
-		proof, err := ProveLookupVector(srs, f, lookup)
+		proof, err := ProveLookupVector(srs, fvector, lookupVector)
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -36,48 +36,45 @@ func TestH(t *testing.T) {
 		}
 	}
 
-	// wrong proof
-	{
-		f[0].SetRandom()
-		proof, err := ProveLookupVector(srs, f, lookup)
-		if err != nil {
-			t.Fatal(err)
-		}
+	// wrong proofs vector
+	// {
+	// 	fvector[0].SetRandom()
 
-		err = VerifyLookupVector(srs, proof)
-		if err == nil {
-			t.Fatal(err)
-		}
-	}
+	// 	proof, err := ProveLookupVector(srs, fvector, lookupVector)
+	// 	if err != nil {
+	// 		t.Fatal(err)
+	// 	}
+
+	// 	err = VerifyLookupVector(srs, proof)
+	// 	if err == nil {
+	// 		t.Fatal(err)
+	// 	}
+	// }
+
+	// lookupTable := make([]Table, 3)
+	// fTable := make([]Table, 3)
+	// for i := 0; i < 3; i++ {
+	// 	lookupTable[i] = make(Table, 8)
+	// 	fTable[i] = make(Table, 7)
+	// 	for j := 0; j < 8; j++ {
+	// 		lookupTable[i][j].SetUint64(uint64(2*i + j))
+	// 	}
+	// 	for j := 0; j < 7; j++ {
+	// 		fTable[i][j].Set(&lookupTable[i][(4*j+1)%8])
+	// 	}
+	// }
+
+	// // correct proof
+	// {
+	// 	proof, err := ProveLookupTables(srs, fTable, lookupTable)
+	// 	if err != nil {
+	// 		t.Fatal(err)
+	// 	}
+
+	// 	err = VerifyLookupTables(srs, proof)
+	// 	if err != nil {
+	// 		t.Fatal(err)
+	// 	}
+	// }
 
 }
-
-// func TestSortByT(t *testing.T) {
-
-// 	a := make(Table, 8)
-// 	b := make(Table, 7)
-// 	for i := 0; i < 8; i++ {
-// 		a[i].SetRandom()
-// 	}
-// 	for i := 0; i < 7; i++ {
-// 		b[i].Set(&a[(4*i)%8])
-// 	}
-// 	sort.Sort(a)
-// 	sort.Sort(b)
-
-// 	for i := 0; i < 8; i++ {
-// 		fmt.Printf("%s\n", a[i].String())
-// 	}
-// 	fmt.Printf("--\n")
-// 	for i := 0; i < 7; i++ {
-// 		fmt.Printf("%s\n", b[i].String())
-// 	}
-
-// 	v := sortByt(b, a)
-// 	fmt.Printf("--\n")
-// 	fmt.Printf("len(v): %d\n", len(v))
-// 	for i := 0; i < len(v); i++ {
-// 		fmt.Printf("%s\n", v[i].String())
-// 	}
-
-// }
